refact: extracting code from router
