+++ api/src/lib/http.js
@@ -1,0 +1,5 @@
+export const HTTP_STATUSES = {
+  OK: 200,
+  CREATED: 201,
+  BAD_REQUEST: 400,
+};
